Consistently use uint16 for ports in diego
[#86337946]

+++ auction_cell_rep/auction_cell_rep.go
@@ -236,7 +236,7 @@ func (a *AuctionCellRep) tasksToContainers(tasks []models.Task) []executor.Conta
 	return containers
 }
 
-func (a *AuctionCellRep) convertPortMappings(containerPorts []uint32) []executor.PortMapping {
+func (a *AuctionCellRep) convertPortMappings(containerPorts []uint16) []executor.PortMapping {
 	out := []executor.PortMapping{}
 	for _, port := range containerPorts {
 		out = append(out, executor.PortMapping{
